% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longestRun.R
\name{longestRun}
\alias{longestRun}
\title{Find the longest run of holiday given a budget of leave days}
\usage{
longestRun(startDate, endDate, daysLeave)
}
\arguments{
\item{startDate}{Date or international format string. Beginning of your search range.}

\item{endDate}{Date or international format string. End of your search range.}

\item{daysLeave}{How many days leave you want to use.}
}
\value{
A list with window (the length of holiday) and the available start dates.
}
\description{
Find the longest run of holiday given a budget of leave days
}
\examples{
longestRun("2017-01-01", "2018-01-10", 10)
longestRun("2017-01-01", "2018-01-10", 3)
}
