% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minDaysLeave.R
\name{minDaysLeave}
\alias{minDaysLeave}
\title{Find the best time to take a fixed length holiday}
\usage{
minDaysLeave(startDate, endDate, window)
}
\arguments{
\item{startDate}{Date or international format string. Beginning of your search range.}

\item{endDate}{Date or international format string. End of your search range.}

\item{window}{The number of consecutive days holiday you would like.
E.g. 3 weeks would be 21 (or 23 if you want an extra weekend)}
}
\value{
A list with the minimum days needed and a vector of potential start dates
}
\description{
Find the best time to take a fixed length holiday
}
\examples{
minDaysLeave("2017-01-01", "2018-01-10", 23)
}
