% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{normalMean}
\alias{normalMean}
\title{Mean of Normals}
\usage{
normalMean(samples = 10)
}
\arguments{
\item{samples}{integer}
}
\description{
Given a number of samples calculate the mean of
the standard normals
}
